requirements.txt
Flask==3.0.0
gunicorn==20.1.0
google-api-core==2.18.0
google-cloud-pubsub==2.21.0
google-cloud-storage==2.16.0
